chore: Fix clippy warnings and documentation discrepancies (Phase 3)
Code quality improvements:
- Fix 375+ clippy warnings across 85+ source files
- Add type aliases for complex function types
- Add #[allow()] attributes for intentional patterns
- Rename from_str methods to parse to avoid trait conflicts
- Fix async lock holding warnings
- Add manual Debug implementations where needed

Documentation fixes:
- Update binary paths in DEPLOYMENT_RUNBOOK.md
- Add /gsp/ prefix to GSP endpoints in API_ENDPOINTS.md
- Fix socket pattern description in CONSENSUS.md
- Correct tweak computation description in GHOST_KEYS.md
- Update constants in NODE_CAPABILITIES.md

+++ bins/ghost-node-tui/src/pages/l2_service.rs
@@ -162,7 +162,7 @@ fn render_epoch_progress(f: &mut Frame, area: Rect, app: &App) {
             lines.push(Line::from(vec![
                 Span::styled("Next reconciliation: ", Style::default().fg(Color::Gray)),
                 Span::styled(
-                    format_duration(remaining_secs as u64),
+                    format_duration(remaining_secs),
                     Style::default().fg(Color::Yellow),
                 ),
             ]));
